Enhanced bearer authentication, added and splitted routers in portions
